+++ src/ic_registry/model/mod.rs
@@ -1,0 +1,13 @@
+mod data_center;
+mod node;
+mod node_operator;
+mod node_provider;
+mod registry;
+mod request;
+
+pub use data_center::{MainnetDataCenter, MainnetDataCenterList};
+pub use node::{MainnetNode, MainnetNodeList};
+pub use node_operator::{MainnetNodeOperator, MainnetNodeOperatorList};
+pub use node_provider::{MainnetNodeProvider, MainnetNodeProviderList};
+pub use registry::MainnetRegistryVersion;
+pub use request::MainnetRegistryFetchRequest;