+++ src/ic_registry/source/mod.rs
@@ -1,0 +1,13 @@
+mod agent;
+mod governance;
+mod nodes;
+mod registry;
+mod subnet_catalog;
+
+pub(super) use governance::fetch_mainnet_node_provider_list_async;
+pub(super) use nodes::{
+    fetch_mainnet_data_center_list_async, fetch_mainnet_node_list_async,
+    fetch_mainnet_node_operator_list_async,
+};
+pub(super) use registry::fetch_mainnet_registry_version_async;
+pub(super) use subnet_catalog::fetch_mainnet_subnet_catalog_async;